refactor(datapage): split IndicatorBrief into two components
