+++ site/owid.scss
@@ -27,7 +27,8 @@
 
 @import "../packages/@ourworldindata/components/src/CodeSnippet/code-snippet.scss";
 @import "../packages/@ourworldindata/components/src/ExpandableToggle/ExpandableToggle.scss";
-@import "../packages/@ourworldindata/components/src/IndicatorBrief/IndicatorBrief.scss";
+@import "../packages/@ourworldindata/components/src/IndicatorKeyData/IndicatorKeyData.scss";
+@import "../packages/@ourworldindata/components/src/IndicatorKeyDescription/IndicatorKeyDescription.scss";
 @import "../packages/@ourworldindata/components/src/IndicatorSources/IndicatorSources.scss";
 @import "../packages/@ourworldindata/components/src/IndicatorProcessing/IndicatorProcessing.scss";
